feat(onitama): add mystic swap power functionality
Implement the mystic swap power (product ID 12) that allows players to swap their master with a student piece. The power can only be used when the player has at least one student piece on the board. Also update the UI to display the new power's icon (🔄) and appropriate status messages.